Fix race condition when using external authentication
Currently, when authenticating using the "external" flow, a socket is
bound to listen for incoming connections from the browser, which POSTs
back the authentication token obtained from the SAML flow. A race
condition occurs where some browsers will also immediately issues
requests on multiple connections to the same port to try and fetch
common web assets, such as the favicon.ico. These multiple requests to
the socket cause certain vars to be overridden, leading to a situation
where authentication can hang.

Concretely, the expected POST request can be immediately followed by one
or more favion.ico requests, which overwrite a previously populated
encodedSamlResponse var with an empty string, resulting the for-loop
hanging, waiting for new connections.

Fix the race condition by setting up a goroutine that will poll the
socket for new connections, passing each into a channel that the main
thread handles synchronously in a for loop on the channel. The channel
is closed upon receipt of a successful SAML response, or an error on the
socket.